pyproject.toml
[tool.poetry]
name = "email-bot"
version = "0.1.0"
description = ""
authors = ["Alexandr Svetlov <s3t1qs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1"
django-ninja = "^0.22.2"
psycopg2-binary = "^2.9.7"
redis = "^5.0.0"
celery = "^5.3.4"
aiogram = "2.14.3"
uvicorn = "^0.23.2"
flower = "^2.0.1"
aioimaplib = "^1.0.1"
bs4 = "^0.0.1"
chardet = "^5.2.0"
types-chardet = "^5.0.4.6"
pillow = "^10.0.1"
django-redis = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
